fix: restrict session title route exports
