+++ packages/web/src/app/api/sessions/[id]/title/parse-request.ts
@@ -1,0 +1,8 @@
+export function parseSessionTitlePatchBody(body: unknown): { title?: string } | null {
+  if (!body || typeof body !== "object" || Array.isArray(body)) return null;
+
+  const title = (body as { title?: unknown }).title;
+  if (title !== undefined && typeof title !== "string") return null;
+
+  return { title };
+}
