+++ packages/web/src/app/api/sessions/[id]/title/route.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { parseSessionTitlePatchBody } from "./route";
+import { parseSessionTitlePatchBody } from "./parse-request";
 
 describe("session title API route", () => {
   describe("parseSessionTitlePatchBody", () => {
